[plugin.video.cheddar] 1.0.1
